WebVisualizer: Display_info fix in visualization.
In web_visualizer, wrt display_info we were doing json loads of the meta data received
from message bus which is not required. Earlier we were doing json dumps before sending
the data over msg_bus, but now as we are not doing json dumps before sending, json loads
on the receiver side is not needed.

Signed-off-by: sagartp <sagar.tp@intel.com>